refactor: remove index.jsx pattern and standardize naming conventions - Removed all index.jsx files in component folders - Changed folder names from CamelCase to kebab-case - Renamed component files to use PascalCase explicitly - Updated all import statements to reference explicit file names - Eliminated the index.jsx pattern throughout the codebase
